Added Dashboard order list with show details and delete

+++ frontend/src/screens/ProfileScreen.js
@@ -1,25 +1,12 @@
+/* eslint-disable no-restricted-globals */
+/* eslint-disable no-alert */
 /* eslint-disable func-names */
 /* eslint-disable object-shorthand */
 // import { update } from '../api.js';
 import { updateSpring, getMyOrdersSpring, deleteOrderSpring } from "../api_spring.js";
 import { clearUser, getUserInfo, setUserInfo } from '../localStorage.js';
-import { showLoading, hideLoading, showMessage, redirectUser, reRender } from '../utils.js';
+import { showLoading, hideLoading, showMessage, redirectUser, reRender, printDate, printTime  } from '../utils.js';
 
-const PrintDate = (t) => {
-// 2021-11-16T04:57:17.118+00:00
-  const date = t.slice(8,10)
-  const month = t.slice(5,7)
-  const year = t.slice(0,4)
-  return `${date}/${month}/${year}`;
-}
-
-const PrintTime = (t) => {
-  
-  const temp = parseInt(t.slice(11, 14), 10) + 8
-  const hour = `${temp >= 24 ? temp-24 : temp}`;
-  const minute = t.slice(14, 16);
-  return `${hour}:${minute}`;
-}
 
 const ProfileScreen = {
 
@@ -56,13 +43,21 @@ const ProfileScreen = {
       this.orders.forEach(order=>{
         const orderId = document.getElementById(`delete-order-${order._id}`)
         orderId.addEventListener('click', async ()=>{
-          console.log(order._id)
+          if (confirm('Are you sure to delete this order?')) {
           const { message } = await deleteOrderSpring(order._id)
           console.log(message)
           showMessage(message)
           reRender(ProfileScreen)
+          }
         })
       })
+      this.orders.forEach((order) => {
+        const orderId = document.getElementById(`${order._id}`);
+        orderId.addEventListener("click", async () => {
+          document.location.hash = `/order/${order._id}`;
+          }
+        );
+      });
       
   },
 
@@ -127,17 +122,23 @@ const ProfileScreen = {
                       (order) => `
           <tr>
             <td>${order._id}</td>
-            <td>${PrintDate(order.createdAt)}</td>
-            <td>${PrintTime(order.createdAt)}</td>
+            <td>${printDate(order.createdAt)}</td>
+            <td>${printTime(order.createdAt)}</td>
             <td>$${order.totalPrice}</td>
             <!--td>${order.paidAt || "No"}</td-->
             <!--td>${order.deliveryAt || "No"}</td-->
-            <td><a href="/#/order/${order._id}">DETAILS</a> 
-                <span id="delete-order-${order._id}" class="px-6 cursor-pointer hover:text-yellow-500">DELETE</span>
+            <!--td><a href="/#/order/${order._id}">DETAILS</a> 
+                <span id="delete-order-${
+                  order._id
+                }" class="px-6 cursor-pointer hover:text-yellow-500">DELETE</span>
+            </td-->
+            <td>
+            <button id="${order._id}" class="details-button">Details</button>
+            <button id="delete-order-${order._id}" class="delete-button">Delete</button>
             </td>
           </tr>
           `).join("\n")
-            }
+          }
           </tbody>
         </table>
       </div>
